Reorganize app module and refactore CLI
