+++ Cargo.toml
@@ -14,8 +14,14 @@ description = """
 
 default-run = "basecoin"
 
+
+[features]
+default = []
+tower-abci = ["dep:tower-abci", "dep:tower"]
+
 [dependencies]
 base64 = { version = "0.21", default-features = false, features = ["alloc"] }
+clap = { version = "4.3.0", features = ["derive", "wrap_help"] }
 cosmrs = "0.11.0"
 displaydoc = { version = "0.2", default-features = false }
 derive_more = { version = "0.99.17", default-features = false, features = ["from", "into", "display"] }
@@ -26,19 +32,21 @@ ics23 = { version = "=0.9.0", default-features = false }
 primitive-types = { version = "0.12.0", default-features = false, features = ["serde_no_std"] }
 prost = { version = "0.11.6", default-features = false }
 serde = "1.0"
+serde_derive = { version = "1.0.104", default-features = false }
 serde_json = "1.0"
 sha2 = "0.10.2"
-structopt = "0.3.21"
 tendermint = "=0.31.1"
 tendermint-abci = "=0.31.1"
 tendermint-proto = "=0.31.1"
+tendermint-rpc = {version = "0.31.1", features = ["http-client"] }
+toml = "0.7"
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
 tonic = "0.9"
 tracing = "0.1.26"
 tracing-subscriber = "0.3.16"
 
-tower-abci = "0.7"
-tower = { version = "0.4", features = ["full"]}
+tower-abci = { version = "0.7", optional = true }
+tower = { version = "0.4", features = ["full"], optional = true }
 
 [dev-dependencies]
 bytes = "1.0.1"
