+++ src/app/interface/mod.rs
@@ -1,0 +1,7 @@
+//! Different interface implementations to interact with the underlying
+//! consensus engine.
+
+pub mod tendermint;
+
+#[cfg(feature = "tower-abci")]
+pub mod tower_abci;
