+++ src/cli/mod.rs
@@ -1,1 +1,1 @@
-pub mod option;
+pub mod command;
