+++ src/lib.rs
@@ -9,6 +9,7 @@
 #![forbid(unsafe_code)]
 pub mod app;
 pub mod cli;
+pub mod config;
 pub mod error;
 mod helper;
 pub mod modules;
