+++ src/modules/upgrade/mod.rs
@@ -1,3 +1,4 @@
 pub mod impls;
 pub mod path;
+pub mod query;
 pub mod service;
